+++ obspy/core/tests/test_event.py
@@ -8,19 +8,17 @@ import builtins
 import copy
 import io
 import os
-import sys
 import unittest
 import warnings
-import tempfile
 
 from matplotlib import rcParams
 import numpy as np
 
-from obspy.core.event import (Catalog, Comment, CreationInfo, Event, Origin,
-                              Pick, ResourceIdentifier, WaveformStreamID,
-                              read_events, Magnitude, FocalMechanism, Arrival)
+from obspy import UTCDateTime, read_events
+from obspy.core.event import (Catalog, Comment, CreationInfo, Event,
+                              FocalMechanism, Magnitude, Origin, Pick,
+                              ResourceIdentifier, WaveformStreamID)
 from obspy.core.event.source import farfield
-from obspy.core.utcdatetime import UTCDateTime
 from obspy.core.util import BASEMAP_VERSION, CARTOPY_VERSION
 from obspy.core.util.base import _get_entry_points
 from obspy.core.util.testing import ImageComparison
@@ -774,378 +772,6 @@ class WaveformStreamIDTestCase(unittest.TestCase):
         self.assertEqual(waveform_id.id, waveform_id.get_seed_string())
 
 
-class ResourceIdentifierTestCase(unittest.TestCase):
-    """
-    Test suite for obspy.core.event.ResourceIdentifier.
-    """
-    def setUp(self):
-        # Clear the Resource Identifier dict for the tests. NEVER do this
-        # otherwise.
-        ResourceIdentifier._ResourceIdentifier__resource_id_weak_dict.clear()
-        # Also clear the tracker.
-        ResourceIdentifier._ResourceIdentifier__resource_id_tracker.clear()
-
-    def test_same_resource_id_different_referred_object(self):
-        """
-        Tests the handling of the case that different ResourceIdentifier
-        instances are created that have the same resource id but different
-        objects. The referred objects should still return the same objects
-        used in the ResourceIdentifier construction or set_referred_object
-        call. However, if an object is set to a resource_id that is not
-        equal to the last object set it should issue a warning.
-        """
-        warnings.simplefilter('default')
-        object_a = UTCDateTime(1000)
-        object_b = UTCDateTime(1000)
-        object_c = UTCDateTime(1001)
-        self.assertFalse(object_a is object_b)
-        id = 'obspy.org/tests/test_resource'
-        res_a = ResourceIdentifier(id=id, referred_object=object_a)
-        # Now create a new resource with the same id but a different object.
-        # This should not raise a warning as the object a and b are equal.
-        with warnings.catch_warnings(record=True) as w:
-            res_b = ResourceIdentifier(id=id, referred_object=object_b)
-            self.assertEqual(len(w), 0)
-        # if the set object is not equal to the last object set to the same
-        # resource_id, however, a warning should be issued.
-        with warnings.catch_warnings(record=True) as w:
-            res_c = ResourceIdentifier(id=id, referred_object=object_c)
-            self.assertEqual(len(w), 1)
-            expected_text = 'which is not equal to the last object bound'
-            self.assertIn(expected_text, str(w[0]))
-        # even though the resource_id are the same, the referred objects
-        # should point to the original (different) objects
-        self.assertIs(object_a, res_a.get_referred_object())
-        self.assertIs(object_b, res_b.get_referred_object())
-        self.assertIs(object_c, res_c.get_referred_object())
-
-    def test_objects_garbage_collection(self):
-        """
-        Test that the ResourceIdentifier class does not mess with the garbage
-        collection of the attached objects.
-        """
-        object_a = UTCDateTime()
-        ref_count = sys.getrefcount(object_a)
-        _res_id = ResourceIdentifier(referred_object=object_a)
-        self.assertEqual(sys.getrefcount(object_a), ref_count)
-        self.assertTrue(bool(_res_id))
-
-    def test_id_without_reference_not_in_global_list(self):
-        """
-        This tests some internal workings of the ResourceIdentifier class.
-        NEVER modify the __resource_id_weak_dict!
-
-        Only those ResourceIdentifiers that have a reference to an object that
-        is referred to somewhere else should stay in the dictionary.
-        """
-        r_dict = ResourceIdentifier._ResourceIdentifier__resource_id_weak_dict
-        _r1 = ResourceIdentifier()  # NOQA
-        self.assertEqual(len(list(r_dict.keys())), 0)
-        # Adding a ResourceIdentifier with an object that does not have a
-        # reference will result in a dict that contains None, but that will
-        # get removed when the resource_id goes out of scope
-        _r2 = ResourceIdentifier(referred_object=UTCDateTime())  # NOQA
-        # raises UserWarning
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", UserWarning)
-            self.assertEqual(_r2.get_referred_object(), None)
-        del _r2  # delete rid to get its id out of r_dict keys
-        # Give it a reference and it will stick around.
-        obj = UTCDateTime()
-        _r3 = ResourceIdentifier(referred_object=obj)  # NOQA
-        self.assertEqual(len(list(r_dict.keys())), 1)
-
-    def test_adding_a_referred_object_after_creation(self):
-        """
-        Check that the referred objects can also be made available after the
-        ResourceIdentifier instances have been created.
-        """
-        obj = UTCDateTime()
-        res_id = "obspy.org/time/test"
-        ref_a = ResourceIdentifier(res_id)
-        ref_b = ResourceIdentifier(res_id)
-        ref_c = ResourceIdentifier(res_id)
-        # All three will have no resource attached.
-        self.assertEqual(ref_a.get_referred_object(), None)
-        self.assertEqual(ref_b.get_referred_object(), None)
-        self.assertEqual(ref_c.get_referred_object(), None)
-        # Setting the object for one will make it available to all other
-        # instances, provided they weren't bound to specific objects.
-        ref_b.set_referred_object(obj)
-        self.assertIs(ref_a.get_referred_object(), obj)
-        self.assertIs(ref_b.get_referred_object(), obj)
-        self.assertIs(ref_c.get_referred_object(), obj)
-
-    def test_getting_gc_no_shared_resource_id(self):
-        """
-        Test that calling get_referred_object on a resource id whose object
-        has been garbage collected, and whose resource_id is unique,
-        returns None
-        """
-        obj1 = UTCDateTime()
-        rid1 = ResourceIdentifier(referred_object=obj1)
-        # delete obj1, make sure rid1 return None
-        del obj1
-        # raises UserWarning
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", UserWarning)
-            self.assertIs(rid1.get_referred_object(), None)
-
-    def test_getting_gc_with_shared_resource_id(self):
-        """
-        Test that calling get_referred_object on a resource id whose object
-        has been garbage collected, but that has another object that shares
-        the same resource_id, returns the other object with the same resource
-        id and issues a warning
-        """
-        uri = 'testuri'
-        obj1 = UTCDateTime(1000)
-        obj2 = UTCDateTime(1000)
-        rid1 = ResourceIdentifier(uri, referred_object=obj1)
-        rid2 = ResourceIdentifier(uri, referred_object=obj2)
-        self.assertFalse(rid1.get_referred_object() is
-                         rid2.get_referred_object())
-        self.assertNotEqual(rid1._object_id, rid2._object_id)
-        del obj1
-        warnings.simplefilter('default')
-        with warnings.catch_warnings(record=True) as w:
-            rid1.get_referred_object()
-            self.assertEqual(len(w), 1)
-            self.assertIn('The object with identity', str(w[0]))
-        # now both rids should return the same object
-        self.assertIs(rid1.get_referred_object(), rid2.get_referred_object())
-        # the object id should now be bound to obj2
-        self.assertEqual(rid1._object_id, rid2._object_id)
-
-    def test_resources_in_global_dict_get_garbage_collected(self):
-        """
-        Tests that the ResourceIdentifiers in the class level resource dict get
-        deleted if they have no other reference and the object they refer to
-        goes out of scope.
-        """
-        obj_a = UTCDateTime()
-        obj_b = UTCDateTime()
-        res1 = ResourceIdentifier(referred_object=obj_a)
-        res2 = ResourceIdentifier(referred_object=obj_b)
-        # Now two keys should be in the global dict.
-        rdict = ResourceIdentifier._ResourceIdentifier__resource_id_weak_dict
-        self.assertEqual(len(list(rdict.keys())), 2)
-        del obj_a, obj_b
-        # raises UserWarnings
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", UserWarning)
-            self.assertIs(res1.get_referred_object(), None)
-            self.assertIs(res2.get_referred_object(), None)
-
-    def test_quakeml_regex(self):
-        """
-        Tests that regex used to check for QuakeML validatity actually works.
-        """
-        # This one contains all valid characters. It should pass the
-        # validation.
-        res_id = (
-            "smi:abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-            "1234567890-.*()_~'/abcdefghijklmnopqrstuvwxyzABCDEFGHIKLMNOPQR"
-            "STUVWXYZ0123456789-.*()_~'+?=,;&")
-        res = ResourceIdentifier(res_id)
-        self.assertEqual(res_id, res.get_quakeml_uri())
-        # The id has to valid from start to end. Due to the spaces this cannot
-        # automatically be converted to a correct one.
-        res_id = ("something_before smi:local/something  something_after")
-        res = ResourceIdentifier(res_id)
-        self.assertRaises(ValueError, res.get_quakeml_uri)
-        # A colon is an invalid character.
-        res_id = ("smi:local/hello:yea")
-        res = ResourceIdentifier(res_id)
-        self.assertRaises(ValueError, res.get_quakeml_uri)
-        # Space as well
-        res_id = ("smi:local/hello yea")
-        res = ResourceIdentifier(res_id)
-        self.assertRaises(ValueError, res.get_quakeml_uri)
-        # Dots are fine
-        res_id = ("smi:local/hello....yea")
-        res = ResourceIdentifier(res_id)
-        self.assertEqual(res_id, res.get_quakeml_uri())
-        # Hats not
-        res_id = ("smi:local/hello^^yea")
-        res = ResourceIdentifier(res_id)
-        self.assertRaises(ValueError, res.get_quakeml_uri)
-
-    def test_resource_id_valid_quakemluri(self):
-        """
-        Test that a resource identifier per default (i.e. no arguments to
-        __init__()) gets set up with a QUAKEML conform ID.
-        """
-        rid = ResourceIdentifier()
-        self.assertEqual(rid.id, rid.get_quakeml_uri())
-
-    def test_resource_id_tracking(self):
-        """
-        The class keeps track of all instances.
-        """
-        # Create a couple of lightweight objects for testing purposes.
-        t1 = UTCDateTime(2013, 1, 1)
-        t2 = UTCDateTime(2013, 1, 2)
-        t3 = UTCDateTime(2013, 1, 3)
-
-        # First assert, that all ResourceIds are tracked correctly.
-        r1 = ResourceIdentifier("a", referred_object=t1)
-        r2 = ResourceIdentifier("b", referred_object=t2)
-        r3 = ResourceIdentifier("c", referred_object=t3)
-
-        self.assertEqual(
-            ResourceIdentifier._ResourceIdentifier__resource_id_tracker,
-            {"a": 1, "b": 1, "c": 1})
-
-        # Create a new instance, similar to the first one.
-        r4 = ResourceIdentifier("a", referred_object=t1)
-        self.assertEqual(
-            ResourceIdentifier._ResourceIdentifier__resource_id_tracker,
-            {"a": 2, "b": 1, "c": 1})
-
-        # Now delete r2 and r4. They should not be tracked anymore.
-        del r2
-        del r4
-        self.assertEqual(
-            ResourceIdentifier._ResourceIdentifier__resource_id_tracker,
-            {"a": 1, "c": 1})
-
-        # Delete the two others. Nothing should be tracked any more.
-        del r1
-        del r3
-        self.assertEqual(
-            ResourceIdentifier._ResourceIdentifier__resource_id_tracker, {})
-
-    def test_automatic_dereferring_if_resource_id_goes_out_of_scope(self):
-        """
-        Tests that objects that have no more referrer are no longer stored in
-        the reference dictionary.
-        """
-        t1 = UTCDateTime(2010, 1, 1)  # test object
-        r_dict = ResourceIdentifier._ResourceIdentifier__resource_id_weak_dict
-        rid = 'a'  # test resource id
-
-        # Create object and assert the reference has been created.
-        r1 = ResourceIdentifier(rid, referred_object=t1)
-        self.assertEqual(r1.get_referred_object(), t1)
-        self.assertTrue(rid in r_dict)
-        # Deleting the object should remove the reference.
-        del r1
-        self.assertFalse(rid in r_dict)
-        # Now create two equal references.
-        r1 = ResourceIdentifier(rid, referred_object=t1)
-        r2 = ResourceIdentifier(rid, referred_object=t1)
-        self.assertEqual(r1.get_referred_object(), t1)
-        # Deleting one should not remove the reference.
-        del r1
-        self.assertEqual(r2.get_referred_object(), t1)
-        self.assertTrue(rid in r_dict)
-        # Deleting the second one should
-        del r2
-        self.assertFalse(rid in r_dict)
-
-    def test_initialize_with_resource_identifier(self):
-        """
-        Test initializing an ResourceIdentifier with an ResourceIdentifier.
-        """
-        rid = ResourceIdentifier()
-        rid2 = ResourceIdentifier(str(rid))
-        rid3 = ResourceIdentifier(rid)
-        self.assertEqual(rid, rid2)
-        self.assertEqual(rid, rid3)
-
-    def test_error_message_for_failing_quakeml_id_conversion(self):
-        """
-        Converting an id to a QuakeML compatible id might fail. Test the
-        error message.
-        """
-        invalid_id = "http://example.org"
-        rid = ResourceIdentifier(invalid_id)
-        with self.assertRaises(ValueError) as e:
-            rid.get_quakeml_uri()
-        self.assertEqual(
-            e.exception.args[0],
-            "The id 'http://example.org' is not a valid QuakeML resource "
-            "identifier. ObsPy tried modifying it to "
-            "'smi:local/http://example.org' but it is still not valid. Please "
-            "make sure all resource ids are either valid or can be made valid "
-            "by prefixing them with 'smi:<authority_id>/'. Valid ids are "
-            "specified in the QuakeML manual section 3.1 and in particular "
-            "exclude colons for the final part.")
-
-
-class ResourceIDEventScopeTestCase(unittest.TestCase):
-    """
-    Test suit for ensuring event scoping of objects bound to
-    ResourceIdentifier instances
-    """
-
-    def make_test_catalog(self):
-        """
-        Make a test catalog with fixed resource IDs some of which reference
-        other objects belonging to the event (eg arrivals -> picks)
-        """
-        pick_rid = ResourceIdentifier(id='obspy.org/tests/test_pick')
-        origin_rid = ResourceIdentifier(id='obspy.org/tests/test_origin')
-        arrival_rid = ResourceIdentifier(id='obspy.org/tests/test_arrival')
-        ar_pick_rid = ResourceIdentifier(id='obspy.org/tests/test_pick')
-        catatlog_rid = ResourceIdentifier(id='obspy.org/tests/test_catalog')
-
-        picks = [Pick(time=UTCDateTime(), resource_id=pick_rid)]
-        arrivals = [Arrival(resource_id=arrival_rid, pick_id=ar_pick_rid)]
-        origins = [Origin(arrivals=arrivals, resource_id=origin_rid)]
-        events = [Event(picks=picks, origins=origins)]
-        events[0].preferred_origin_id = str(origin_rid.id)
-        catalog = Catalog(events=events, resource_id=catatlog_rid)
-        # next bind all unbound resource_ids to the current event scope
-        catalog.resource_id.bind_resource_ids()
-        return catalog
-
-    def setUp(self):
-        # Clear the Resource Identifier dict for the tests. NEVER do this
-        # otherwise.
-        ResourceIdentifier._ResourceIdentifier__resource_id_weak_dict.clear()
-        # Also clear the tracker.
-        ResourceIdentifier._ResourceIdentifier__resource_id_tracker.clear()
-        # set the test catalog as an attr for test access
-        self.catalog = self.make_test_catalog()
-        # save the catalog to a temp file for testing reading in the catalog
-        self.catalog_path = tempfile.mkstemp()[1]
-        self.catalog.write(self.catalog_path, 'quakeml')
-        # create a list of equal catalogs/events created with read and copy
-        self.event_list = [
-            self.catalog[0],
-            read_events(self.catalog_path)[0],
-            read_events(self.catalog_path)[0],
-            self.catalog.copy()[0],
-            self.catalog.copy()[0],
-        ]
-
-    def test_preferred_origins(self):
-        """
-        test that the objects bound to the preferred origins are event scoped
-        """
-        for ev in self.event_list:
-            self.assertIs(ev.preferred_origin(), ev.origins[0])
-
-    def test_arrivals_refer_to_picks_in_same_event(self):
-        """
-        ensure the pick_ids of the arrivals refer to the pick belonging
-        to the same event
-        """
-        for ev in self.event_list:
-            pick_id = ev.picks[0].resource_id
-            arrival_pick_id = ev.origins[0].arrivals[0].pick_id
-            self.assertEqual(pick_id, arrival_pick_id)
-            pick = ev.picks[0]
-            arrival_pick = arrival_pick_id.get_referred_object()
-            self.assertIs(pick, arrival_pick)
-
-    def tearDown(self):
-        # remove the temp file
-        os.remove(self.catalog_path)
-
-
 class BaseTestCase(unittest.TestCase):
     """
     Test suite for obspy.core.event.base.
@@ -1230,7 +856,7 @@ class BaseTestCase(unittest.TestCase):
 
         del t1
 
-        self.assertEqual(rid.get_referred_object(), t2)
+        self.assertIs(rid.get_referred_object(), t2)
 
 
 def suite():
@@ -1241,7 +867,6 @@ def suite():
     suite.addTest(unittest.makeSuite(EventTestCase, 'test'))
     suite.addTest(unittest.makeSuite(OriginTestCase, 'test'))
     suite.addTest(unittest.makeSuite(WaveformStreamIDTestCase, 'test'))
-    suite.addTest(unittest.makeSuite(ResourceIdentifierTestCase, 'test'))
     suite.addTest(unittest.makeSuite(BaseTestCase, 'test'))
     return suite
